Version support for warehouse schema gen (#158)
Add default version for wh schema
