+++ v0/destinations/rs/transform.js
@@ -1,11 +1,12 @@
-const { processWarehouseMessage } = require("../util");
+const { processWarehouseMessage } = require("../../../warehouse");
 
-function processSingleMessage(message, destination) {
-  return processWarehouseMessage("rs", message);
+function processSingleMessage(message, schemaVersion) {
+  return processWarehouseMessage("rs", message, schemaVersion);
 }
 
 function process(event) {
-  return processSingleMessage(event.message, event.destination);
+  const schemaVersion = event.request.query.whSchemaVersion || "v1";
+  return processSingleMessage(event.message, schemaVersion);
 }
 
 exports.process = process;
